refactor: main의 Lotto.java ticket 파일로 이동
